russimp: Introduce FileSystem loading
When loading resources from memory buffers, sometimes the resources in
question link to secondary files that must also be loaded.  An example
of this is the loading of .obj files, which store their materials in a
secondary .mtl file.  This file can't be passed in using the
from_buffer() loading path, resulting in the need to load the files from
the host native filesystem using the from_file() loading path.

This change implements a rust safe loading path to allowing applications
to implement their own virtual filesystem interface with assimp. This is
convenient for use in combination with an application specific resource
loader.

To use, callers simply implement the fs::FileSystem trait, which
neccesitates the implementation of open() which returns their own type
as a trait object implementing fs::FileOperations.

This new fs module handles all the unsafe glue required to interface
this safe trait-based interface with the underlying unsafe aiFileIO
parts.